+++ relay-server/src/metrics_extraction/transactions.rs
@@ -1,19 +1,17 @@
+use crate::metrics_extraction::conditional_tagging::run_conditional_tagging;
+use crate::metrics_extraction::utils;
+use crate::metrics_extraction::TaggingRule;
+use relay_common::{SpanStatus, UnixTimestamp};
+use relay_general::protocol::TraceContext;
+use relay_general::protocol::{AsPair, Timestamp, TransactionSource};
+use relay_general::protocol::{Context, ContextInner};
+use relay_general::protocol::{Event, EventType};
+use relay_general::store;
+use relay_general::types::Annotated;
+use relay_metrics::{DurationUnit, Metric, MetricNamespace, MetricUnit, MetricValue};
 use serde::{Deserialize, Serialize};
 use std::collections::{BTreeMap, BTreeSet};
-
-#[cfg(feature = "processing")]
-use {
-    crate::metrics_extraction::conditional_tagging::run_conditional_tagging,
-    crate::metrics_extraction::{utils, TaggingRule},
-    relay_common::{SpanStatus, UnixTimestamp},
-    relay_general::protocol::TraceContext,
-    relay_general::protocol::{AsPair, Event, EventType, Timestamp},
-    relay_general::protocol::{Context, ContextInner},
-    relay_general::store,
-    relay_general::types::Annotated,
-    relay_metrics::{DurationUnit, Metric, MetricNamespace, MetricUnit, MetricValue},
-    std::fmt,
-};
+use std::fmt;
 
 /// The metric on which the user satisfaction threshold is applied.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -51,20 +49,50 @@ pub struct CustomMeasurementConfig {
     limit: usize,
 }
 
+/// Maximum supported version of metrics extraction from transactions.
+///
+/// The version is an integer scalar, incremented by one on each new version.
+const EXTRACT_MAX_VERSION: u16 = 1;
+
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub enum AcceptTransactionNames {
+    /// For some SDKs, accept all transaction names, while for others, apply strict rules.
+    ClientBased,
+
+    /// Only accept transaction names with a low-cardinality source.
+    /// Any value other than "clientBased" will be interpreted as "strict".
+    #[serde(other)]
+    Strict,
+}
+
+impl Default for AcceptTransactionNames {
+    fn default() -> Self {
+        Self::Strict
+    }
+}
+
 /// Configuration for extracting metrics from transaction payloads.
 #[derive(Default, Debug, Clone, Serialize, Deserialize)]
 #[serde(default, rename_all = "camelCase")]
 pub struct TransactionMetricsConfig {
+    /// The required version to extract transaction metrics.
+    version: u16,
     extract_metrics: BTreeSet<String>,
     extract_custom_tags: BTreeSet<String>,
     satisfaction_thresholds: Option<SatisfactionConfig>,
     custom_measurements: CustomMeasurementConfig,
+    accept_transaction_names: AcceptTransactionNames,
 }
 
-#[cfg(feature = "processing")]
+impl TransactionMetricsConfig {
+    pub fn is_enabled(&self) -> bool {
+        self.version > 0 && self.version <= EXTRACT_MAX_VERSION
+    }
+}
+
 const METRIC_NAMESPACE: MetricNamespace = MetricNamespace::Transactions;
 
-#[cfg(feature = "processing")]
 fn get_trace_context(event: &Event) -> Option<&TraceContext> {
     let contexts = event.contexts.value()?;
     let trace = contexts.get("trace").map(Annotated::value);
@@ -77,18 +105,15 @@ fn get_trace_context(event: &Event) -> Option<&TraceContext> {
 
 /// Extract transaction status, defaulting to [`SpanStatus::Unknown`].
 /// Must be consistent with `process_trace_context` in [`relay_general::store`].
-#[cfg(feature = "processing")]
 fn extract_transaction_status(trace_context: &TraceContext) -> SpanStatus {
     *trace_context.status.value().unwrap_or(&SpanStatus::Unknown)
 }
 
-#[cfg(feature = "processing")]
 fn extract_transaction_op(trace_context: &TraceContext) -> Option<String> {
     let op = trace_context.op.value()?;
     Some(op.to_string())
 }
 
-#[cfg(feature = "processing")]
 fn extract_dist(transaction: &Event) -> Option<String> {
     let mut dist = transaction.dist.0.clone();
     store::normalize_dist(&mut dist);
@@ -97,7 +122,6 @@ fn extract_dist(transaction: &Event) -> Option<String> {
 
 /// Extract HTTP method
 /// See <https://github.com/getsentry/snuba/blob/2e038c13a50735d58cc9397a29155ab5422a62e5/snuba/datasets/errors_processor.py#L64-L67>.
-#[cfg(feature = "processing")]
 fn extract_http_method(transaction: &Event) -> Option<String> {
     let request = transaction.request.value()?;
     let method = request.method.value()?;
@@ -106,14 +130,12 @@ fn extract_http_method(transaction: &Event) -> Option<String> {
 
 /// Satisfaction value used for Apdex and User Misery
 /// <https://docs.sentry.io/product/performance/metrics/#apdex>
-#[cfg(feature = "processing")]
 enum UserSatisfaction {
     Satisfied,
     Tolerated,
     Frustrated,
 }
 
-#[cfg(feature = "processing")]
 impl fmt::Display for UserSatisfaction {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
@@ -124,7 +146,6 @@ impl fmt::Display for UserSatisfaction {
     }
 }
 
-#[cfg(feature = "processing")]
 impl UserSatisfaction {
     /// The frustration threshold is always four times the threshold
     /// (see <https://docs.sentry.io/product/performance/metrics/#apdex>)
@@ -143,7 +164,6 @@ impl UserSatisfaction {
 
 /// Extract the the satisfaction value depending on the actual measurement/duration value
 /// and the configured threshold.
-#[cfg(feature = "processing")]
 fn extract_user_satisfaction(
     config: &Option<SatisfactionConfig>,
     transaction: &Event,
@@ -169,11 +189,91 @@ fn extract_user_satisfaction(
     None
 }
 
+fn is_low_cardinality(source: &TransactionSource, treat_unknown_as_low_cardinality: bool) -> bool {
+    match source {
+        // For now, we hope that custom transaction names set by users are low-cardinality.
+        TransactionSource::Custom => true,
+
+        // "url" are raw URLs, potentially containing identifiers.
+        TransactionSource::Url => false,
+
+        // These four are names of software components, which we assume to be low-cardinality.
+        TransactionSource::Route => true,
+        TransactionSource::View => true,
+        TransactionSource::Component => true,
+        TransactionSource::Task => true,
+
+        // "unknown" is the value for old SDKs that do not send a transaction source yet.
+        // Caller decides how to treat this.
+        TransactionSource::Unknown => treat_unknown_as_low_cardinality,
+
+        // Any other value would be an SDK bug, assume high-cardinality and drop.
+        TransactionSource::Other(source) => {
+            relay_log::error!("Invalid transaction source: '{}'", source);
+            false
+        }
+    }
+}
+
+fn is_browser_sdk(event: &Event) -> bool {
+    let sdk_name = event
+        .client_sdk
+        .value()
+        .and_then(|sdk| sdk.name.value())
+        .map(|s| s.as_str())
+        .unwrap_or_default();
+
+    [
+        "sentry.javascript.angular",
+        "sentry.javascript.browser",
+        "sentry.javascript.ember",
+        "sentry.javascript.gatsby",
+        "sentry.javascript.react",
+        "sentry.javascript.remix",
+        "sentry.javascript.vue",
+    ]
+    .contains(&sdk_name)
+}
+
+/// Decide whether we want to keep the transaction name.
+/// High-cardinality sources are excluded to protect our metrics infrastructure.
+/// Note that this will produce a discrepancy between metrics and raw transaction data.
+fn get_transaction_name(
+    event: &Event,
+    accept_transaction_names: &AcceptTransactionNames,
+) -> Option<String> {
+    let original_transaction_name = match event.transaction.value() {
+        Some(name) => name,
+        None => {
+            return None;
+        }
+    };
+
+    // In client-based mode, handling of "unknown" sources depends on the SDK name.
+    // In strict mode, treat "unknown" as high cardinality.
+    let treat_unknown_as_low_cardinality = matches!(
+        accept_transaction_names,
+        AcceptTransactionNames::ClientBased
+    ) && !is_browser_sdk(event);
+
+    let source = event.get_transaction_source();
+    let use_original_name = is_low_cardinality(source, treat_unknown_as_low_cardinality);
+
+    if use_original_name {
+        Some(original_transaction_name.clone())
+    } else {
+        // Pick a sentinel based on the transaction source:
+        match source {
+            TransactionSource::Unknown | TransactionSource::Other(_) => None,
+            _ => Some("<< unparametrized >>".to_owned()),
+        }
+    }
+}
+
 /// These are the tags that are added to all extracted metrics.
-#[cfg(feature = "processing")]
 fn extract_universal_tags(
     event: &Event,
-    custom_tags: &BTreeSet<String>,
+    config: &TransactionMetricsConfig,
 ) -> BTreeMap<String, String> {
     let mut tags = BTreeMap::new();
     if let Some(release) = event.release.as_str() {
@@ -185,8 +285,8 @@ fn extract_universal_tags(
     if let Some(environment) = event.environment.as_str() {
         tags.insert("environment".to_owned(), environment.to_owned());
     }
-    if let Some(transaction) = event.transaction.as_str() {
-        tags.insert("transaction".to_owned(), transaction.to_owned());
+    if let Some(transaction_name) = get_transaction_name(event, &config.accept_transaction_names) {
+        tags.insert("transaction".to_owned(), transaction_name);
     }
 
     // The platform tag should not increase dimensionality in most cases, because most
@@ -210,6 +310,7 @@ fn extract_universal_tags(
         tags.insert("http.method".to_owned(), http_method);
     }
 
+    let custom_tags = &config.extract_custom_tags;
     if !custom_tags.is_empty() {
         // XXX(slow): event tags are a flat array
         if let Some(event_tags) = event.tags.value() {
@@ -233,7 +334,6 @@ fn extract_universal_tags(
 ///
 /// For known measurements, this returns `Some(MetricUnit)`, which can also include
 /// `Some(MetricUnit::None)`. For unknown measurement names, this returns `None`.
-#[cfg(feature = "processing")]
 fn get_metric_measurement_unit(metric: &str) -> Option<MetricUnit> {
     match metric {
         // Web
@@ -262,7 +362,6 @@ fn get_metric_measurement_unit(metric: &str) -> Option<MetricUnit> {
     }
 }
 
-#[cfg(feature = "processing")]
 pub fn extract_transaction_metrics(
     config: &TransactionMetricsConfig,
     breakdowns_config: Option<&store::BreakdownsConfig>,
@@ -302,7 +401,6 @@ pub fn extract_transaction_metrics(
     !added_slice.is_empty()
 }
 
-#[cfg(feature = "processing")]
 fn extract_transaction_metrics_inner(
     config: &TransactionMetricsConfig,
     breakdowns_config: Option<&store::BreakdownsConfig>,
@@ -325,7 +423,7 @@ fn extract_transaction_metrics_inner(
         None => return,
     };
 
-    let tags = extract_universal_tags(event, &config.extract_custom_tags);
+    let tags = extract_universal_tags(event, config);
 
     // Measurements
     if let Some(measurements) = event.measurements.value() {
@@ -436,7 +534,6 @@ fn extract_transaction_metrics_inner(
     }
 }
 
-#[cfg(feature = "processing")]
 fn get_measurement_rating(name: &str, value: f64) -> Option<String> {
     let rate_range = |meh_ceiling: f64, poor_ceiling: f64| {
         debug_assert!(meh_ceiling < poor_ceiling);
@@ -461,10 +558,12 @@ fn get_measurement_rating(name: &str, value: f64) -> Option<String> {
 #[cfg(test)]
 #[cfg(feature = "processing")]
 mod tests {
+
     use super::*;
 
     use crate::metrics_extraction::TaggingRule;
     use insta::assert_debug_snapshot;
+    use relay_general::protocol::{ClientSdkInfo, TransactionInfo};
     use relay_general::store::BreakdownsConfig;
     use relay_general::types::Annotated;
     use relay_metrics::DurationUnit;
@@ -481,6 +580,7 @@ mod tests {
             "dist": "foo ",
             "environment": "fake_environment",
             "transaction": "mytransaction",
+            "transaction_info": {"source": "custom"},
             "user": {
                 "id": "user123"
             },
@@ -741,11 +841,10 @@ mod tests {
             panic!(); // Duration must be set
         }
 
-        assert_eq!(duration_metric.tags.len(), 5);
+        assert_eq!(duration_metric.tags.len(), 4);
         assert_eq!(duration_metric.tags["release"], "1.2.3");
         assert_eq!(duration_metric.tags["transaction.status"], "ok");
         assert_eq!(duration_metric.tags["environment"], "fake_environment");
-        assert_eq!(duration_metric.tags["transaction"], "mytransaction");
         assert_eq!(duration_metric.tags["platform"], "other");
     }
 
@@ -793,12 +892,12 @@ mod tests {
         assert_eq!(metrics.len(), 2);
 
         let duration_metric = &metrics[0];
-        assert_eq!(duration_metric.tags.len(), 4);
+        assert_eq!(duration_metric.tags.len(), 3);
         assert_eq!(duration_metric.tags["satisfaction"], "tolerated");
         assert_eq!(duration_metric.tags["transaction.status"], "ok");
 
         let user_metric = &metrics[1];
-        assert_eq!(user_metric.tags.len(), 4);
+        assert_eq!(user_metric.tags.len(), 3);
         assert_eq!(user_metric.tags["satisfaction"], "tolerated");
     }
 
@@ -845,7 +944,7 @@ mod tests {
         assert_eq!(metrics.len(), 1);
 
         for metric in metrics {
-            assert_eq!(metric.tags.len(), 3);
+            assert_eq!(metric.tags.len(), 2);
             assert_eq!(metric.tags["satisfaction"], "satisfied");
         }
     }
@@ -887,7 +986,7 @@ mod tests {
         assert_eq!(metrics.len(), 1);
 
         for metric in metrics {
-            assert_eq!(metric.tags.len(), 2);
+            assert_eq!(metric.tags.len(), 1);
             assert!(!metric.tags.contains_key("satisfaction"));
         }
     }
@@ -937,7 +1036,6 @@ mod tests {
                     timestamp: UnixTimestamp(1619420402),
                     tags: {
                         "platform": "other",
-                        "transaction": "foo",
                     },
                 },
                 Metric {
@@ -949,7 +1047,6 @@ mod tests {
                     tags: {
                         "measurement_rating": "good",
                         "platform": "other",
-                        "transaction": "foo",
                     },
                 },
                 Metric {
@@ -960,7 +1057,6 @@ mod tests {
                     timestamp: UnixTimestamp(1619420402),
                     tags: {
                         "platform": "other",
-                        "transaction": "foo",
                     },
                 },
             ]
@@ -1039,7 +1135,6 @@ mod tests {
                 {
                     let mut tags = BTreeMap::new();
                     tags.insert("satisfaction".to_owned(), "tolerated".to_owned());
-                    tags.insert("transaction".to_owned(), "foo".to_owned());
                     tags.insert("platform".to_owned(), "other".to_owned());
                     tags
                 }
@@ -1120,7 +1215,6 @@ mod tests {
                     let mut tags = BTreeMap::new();
                     tags.insert("satisfaction".to_owned(), "frustrated".to_owned());
                     tags.insert("measurement_rating".to_owned(), "good".to_owned());
-                    tags.insert("transaction".to_owned(), "foo".to_owned());
                     tags.insert("platform".to_owned(), "other".to_owned());
                     tags
                 }
@@ -1200,5 +1294,177 @@ mod tests {
                 ("platform".to_string(), "other".to_string())
             ])
         );
+    }
+
+    /// Helper function to check if the transaction name is set correctly
+    fn extract_transaction_name(
+        sdk_name: &str,
+        source: &TransactionSource,
+        strategy: AcceptTransactionNames,
+    ) -> Option<String> {
+        let json = r#"
+        {
+            "type": "transaction",
+            "transaction": "foo",
+            "timestamp": "2021-04-26T08:00:00+0100",
+            "start_timestamp": "2021-04-26T07:59:01+0100",
+            "contexts": {"trace": {}}
+        }
+        "#;
+
+        let mut config: TransactionMetricsConfig = serde_json::from_str(
+            r#"
+        {
+            "extractMetrics": [
+                "d:transactions/duration@millisecond"
+            ]
+        }
+        "#,
+        )
+        .unwrap();
+
+        let mut event = Annotated::<Event>::from_json(json).unwrap();
+        {
+            let event = event.value_mut().as_mut().unwrap();
+            event.transaction_info.set_value(Some(TransactionInfo {
+                source: Annotated::new(source.clone()),
+                original: Annotated::empty(),
+            }));
+
+            event.client_sdk.set_value(Some(ClientSdkInfo {
+                name: Annotated::new(sdk_name.to_owned()),
+                ..Default::default()
+            }));
+        }
+
+        config.accept_transaction_names = strategy;
+
+        let mut metrics = vec![];
+        extract_transaction_metrics(&config, None, &[], event.value().unwrap(), &mut metrics);
+
+        assert_eq!(metrics.len(), 1);
+        metrics[0].tags.get("transaction").cloned()
+    }
+
+    #[test]
+    fn test_js_unknown_strict() {
+        let name = extract_transaction_name(
+            "sentry.javascript.browser",
+            &TransactionSource::Unknown,
+            AcceptTransactionNames::Strict,
+        );
+        assert!(name.is_none());
+    }
+
+    #[test]
+    fn test_js_unknown_client_based() {
+        let name = extract_transaction_name(
+            "sentry.javascript.browser",
+            &TransactionSource::Unknown,
+            AcceptTransactionNames::ClientBased,
+        );
+        assert!(name.is_none());
+    }
+
+    #[test]
+    fn test_js_url_strict() {
+        let name = extract_transaction_name(
+            "sentry.javascript.browser",
+            &TransactionSource::Url,
+            AcceptTransactionNames::Strict,
+        );
+        assert_eq!(name, Some("<< unparametrized >>".to_owned()));
+    }
+
+    #[test]
+    fn test_js_url_client_based() {
+        let name = extract_transaction_name(
+            "sentry.javascript.browser",
+            &TransactionSource::Url,
+            AcceptTransactionNames::Strict,
+        );
+        assert_eq!(name, Some("<< unparametrized >>".to_owned()));
+    }
+
+    #[test]
+    fn test_other_client_unknown_strict() {
+        let name = extract_transaction_name(
+            "some_client",
+            &TransactionSource::Unknown,
+            AcceptTransactionNames::Strict,
+        );
+        assert!(name.is_none());
+    }
+
+    #[test]
+    fn test_other_client_unknown_client_based() {
+        let name = extract_transaction_name(
+            "some_client",
+            &TransactionSource::Unknown,
+            AcceptTransactionNames::ClientBased,
+        );
+        assert_eq!(name, Some("foo".to_owned()));
+    }
+
+    #[test]
+    fn test_other_client_url_strict() {
+        let name = extract_transaction_name(
+            "some_client",
+            &TransactionSource::Url,
+            AcceptTransactionNames::Strict,
+        );
+        assert_eq!(name, Some("<< unparametrized >>".to_owned()));
+    }
+
+    #[test]
+    fn test_other_client_url_client_based() {
+        let name = extract_transaction_name(
+            "some_client",
+            &TransactionSource::Url,
+            AcceptTransactionNames::Strict,
+        );
+        assert_eq!(name, Some("<< unparametrized >>".to_owned()));
+    }
+
+    #[test]
+    fn test_any_client_route_strict() {
+        let name = extract_transaction_name(
+            "some_client",
+            &TransactionSource::Route,
+            AcceptTransactionNames::Strict,
+        );
+        assert_eq!(name, Some("foo".to_owned()));
+    }
+
+    #[test]
+    fn test_any_client_route_client_based() {
+        let name = extract_transaction_name(
+            "some_client",
+            &TransactionSource::Route,
+            AcceptTransactionNames::ClientBased,
+        );
+        assert_eq!(name, Some("foo".to_owned()));
+    }
+
+    #[test]
+    fn test_parse_transaction_name_strategy() {
+        for (config, expected_strategy) in [
+            (r#"{}"#, AcceptTransactionNames::Strict),
+            (
+                r#"{"acceptTransactionNames": "unknown-strategy"}"#,
+                AcceptTransactionNames::Strict,
+            ),
+            (
+                r#"{"acceptTransactionNames": "strict"}"#,
+                AcceptTransactionNames::Strict,
+            ),
+            (
+                r#"{"acceptTransactionNames": "clientBased"}"#,
+                AcceptTransactionNames::ClientBased,
+            ),
+        ] {
+            let config: TransactionMetricsConfig = serde_json::from_str(config).unwrap();
+            assert_eq!(config.accept_transaction_names, expected_strategy);
+        }
     }
 }